js/xpconnect/tests/chrome/test_weakmap_keys_preserved.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=673468
-->
<window title="Mozilla Bug "
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id="
     target="_blank">Mozilla Bug 673468</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[
  /** Test for Bug 673468 **/

  let Cc = Components.classes;
  let Cu = Components.utils;
  let Ci = Components.interfaces;

  let system = Cc["@mozilla.org/systemprincipal;1"].createInstance();
  let sandbox = Cu.Sandbox(system);
  let map = sandbox.WeakMap();
  let obj = {};
  map.set(obj, {});

  Cu.forceGC();

  ok(map.has(obj), "Weakmap still contains our wrapper!");
  ]]>
  </script>
</window>
